+++ src/Models/User.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App\Models;
+namespace Skeleton\Models;
 
 /**
  * User Model
  *
- * @package App
+ * @package Skeleton
  * @author  Rougin Royce Gutib <rougingutib@gmail.com>
  */
 class User extends \Illuminate\Database\Eloquent\Model