+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
-All Notable changes to `App` will be documented in this file
+All Notable changes to `Skeleton` will be documented in this file
 
 ## 0.1.0 - CURRENT_DATE
 
 ### Added
-- `App` file and directory structure
+- `Skeleton` file and directory structure